backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
yfinance==0.2.35
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
anthropic==0.18.1
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
pytesseract==0.3.10
pillow==10.2.0
pytest==7.4.4
